# core/docs/swagger.yaml
definitions:
  apprun_modules_auth_service.CreateProjectRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  apprun_modules_auth_service.ProjectResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  apprun_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  apprun_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        $ref: '#/definitions/apprun_pkg_response.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@websoft9.com
    name: API Support
  description: AppRun Platform REST API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AppRun API
  version: "1.0"
paths:
...
  /api/projects:
    get:
      description: Lists all projects where the authenticated user is a member
      produces:
      - application/json
      responses:
        "200":
          description: List of projects
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/apprun_modules_auth_service.ProjectResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: List user's projects
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Creates a new project with the authenticated user as owner
      parameters:
      - description: Project creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apprun_modules_auth_service.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/apprun_modules_auth_service.ProjectResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - Projects
  /api/projects/{id}:
    delete:
      description: Archives a project (soft delete)
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - Projects
    get:
      description: Gets detailed information about a specific project
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project details
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/apprun_modules_auth_service.ProjectResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get project details
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Updates project information (name, description)
      parameters:
      - description: Project UUID
        in: path
        name: id
        required: true
        type: string
      - description: Project update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apprun_modules_auth_service.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/apprun_modules_auth_service.ProjectResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Projects
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"